requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
google-generativeai==0.8.3
python-dotenv==1.0.1
lxml==5.3.0
gunicorn==23.0.0
